+++ src/main/java/racingcar/Game.java
@@ -1,21 +1,19 @@
 package racingcar;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import racingcar.domain.Cars;
-import racingcar.dto.Result;
+import racingcar.controller.GameController;
 import racingcar.strategy.MoveStrategy;
 import racingcar.strategy.NumberOverFourStrategy;
 import racingcar.strategy.NumberStrategy;
 import racingcar.strategy.RandomNumberStrategy;
 import racingcar.view.InputView;
+import racingcar.view.InputViewImpl;
 import racingcar.view.OutputView;
+import racingcar.view.OutputViewImpl;
 
 public class Game {
 
-	private static final InputView INPUT_VIEW = new InputView();
-	private static final OutputView OUTPUT_VIEW = new OutputView();
+	private static final InputView INPUT_VIEW = new InputViewImpl();
+	private static final OutputView OUTPUT_VIEW = new OutputViewImpl();
 	private static final int RANDOM_NUMBER_BOUND_INCLUSIVE = 9;
 
 	public static void main(String[] args) {
